feat(music.js): Use queue functions and database

+++ funcs/music.js
@@ -11,12 +11,13 @@ import path from "path";
 // import LyricsFunctions from "./lyrics.js";
 import LyricsFunctions from "./lyricsGenius.js";
 import MusicDownloader from "./downloader.js";
+import Database from "./database.js";
 
 const readFileAsync = promisify(fs.readFile);
 const writeFileAsync = promisify(fs.writeFile);
 
 class MusicFunctions {
-  constructor(config) {
+  constructor(config, database) {
     this.spotifyClient = new SpotifyGet({
       consumer: {
         key: process.env.SPOTIFY_CLIENT_ID,
@@ -47,6 +48,10 @@ class MusicFunctions {
 
     this.lyricsController = new LyricsFunctions(config);
     this.downloader = new MusicDownloader(config);
+    this.database = new Database(config);
+    // this.database = database;
+
+    this.isQueueDownloading = false;
   }
 
   async downloadCoverTrack(track_data) {
@@ -62,6 +67,9 @@ class MusicFunctions {
           )
           .on("close", () => {
             resolve(fileName);
+          })
+          .on("error", (error) => {
+            reject(error);
           });
       });
     });
@@ -84,28 +92,39 @@ class MusicFunctions {
       `${name} - ${album_artist}.jpg`
     );
 
-    console.log("Tagging", mp3InputPath, "with", coverPath, "to", mp3OutputPath)
+    console.log(
+      "Tagging",
+      mp3InputPath,
+      "with",
+      coverPath,
+      "to",
+      mp3OutputPath
+    );
 
-    const mp3Data = await readFileAsync(mp3InputPath);
-    const imageData = await readFileAsync(coverPath);
-    const tags = id3.read(mp3Data);
+    try {
+      const mp3Data = await readFileAsync(mp3InputPath);
+      const imageData = await readFileAsync(coverPath);
+      const tags = id3.read(mp3Data);
 
-    tags.image = {
-      mime: "image/jpeg",
-      type: {
-        id: 3,
-        name: "front cover",
-      },
-      description: "Cover",
-      imageBuffer: imageData,
-    };
+      tags.image = {
+        mime: "image/jpeg",
+        type: {
+          id: 3,
+          name: "front cover",
+        },
+        description: "Cover",
+        imageBuffer: imageData,
+      };
 
-    const taggedMp3Data = id3.write(tags, mp3Data);
-    await writeFileAsync(mp3OutputPath, taggedMp3Data);
+      const taggedMp3Data = id3.write(tags, mp3Data);
+      await writeFileAsync(mp3OutputPath, taggedMp3Data);
 
-    return mp3OutputPath;
+      return mp3OutputPath;
+    } catch (error) {
+      console.error("Error while tagging", error);
+      return false;
+    }
   }
-
 
   async getArtistTracks(artist_id) {
     const token = await this.spotifyClient.getToken();
@@ -120,8 +139,6 @@ class MusicFunctions {
 
     const data = await response.json();
 
-    // return data.tracks;
-
     data.tracks = data.tracks.map((item) => {
       let artist = item.artists.map((a) => a.name).join(" / ");
 
@@ -130,7 +147,7 @@ class MusicFunctions {
         artist: utils.cleanName(artist),
         album_artist: utils.cleanName(item.artists[0].name),
         cover: item.album.images[0].url,
-        
+
         id: item.id,
         uri: item.external_urls.spotify,
         track_position: item.track_number,
@@ -152,46 +169,8 @@ class MusicFunctions {
           },
           release_date: item.album.release_date,
           year: item.album.release_date.split("-")[0],
-        }
+        },
       };
-
-
-      // let artists = item.artists.map((a) => {
-      //   return {
-      //     name: a.name,
-      //     id: a.i,
-      //     uri: a.external_urls.spotify,
-      //   };
-      // });
-
-      // let album = {
-      //   id: item.album.id,
-      //   uri: item.album.external_urls.spotify,
-      //   name: item.album.name,
-      //   image: {
-      //     url: item.album.images[0].url,
-      //     height: item.album.images[0].height,
-      //     width: item.album.images[0].width,
-      //   },
-      //   release_date: item.album.release_date,
-      //   year: item.album.release_date.split("-")[0],
-      // };
-
-      // let artist = artists.map((a) => a.name).join(" / ");
-
-      // return {
-      //   name: item.name,
-      //   artist,
-      //   album_artist: artists[0].name,
-      //   cover: album.image.url,
-      //   id: item.id,
-      //   uri: item.external_urls.spotify,
-      //   artists,
-      //   duration_ms: item.duration_ms,
-      //   album,
-      //   preview_url: item.preview_url,
-      //   track_position: item.track_number,
-      // };
     });
 
     return data;
@@ -203,7 +182,7 @@ class MusicFunctions {
     if (type == "*") {
       type = "artist,album,track";
     }
-    
+
     const searchDatas = await this.spotifyClient.search({
       q: `${query}`,
       type: type,
@@ -228,7 +207,7 @@ class MusicFunctions {
     }
 
     if (searchDatas.albums?.items) {
-      searchDatas.albums.items = searchDatas.albums.items.map( (item) => {
+      searchDatas.albums.items = searchDatas.albums.items.map((item) => {
         return {
           name: utils.cleanName(item.name),
           id: item.id,
@@ -292,16 +271,54 @@ class MusicFunctions {
     };
   }
 
-  async downloadFromDatas(track_data) {
-    // Download a song from Spotify using the data object returned by the search function
-    let { name, album_artist, album, track_position } = track_data;
+  async downloadNextSongInQueue() {
+    console.log("isQueueDownloading", this.isQueueDownloading);
+    if (this.isQueueDownloading) return;
+    const queueSize = await this.database.getPendingQueueSize();
+    console.log("queueSize", queueSize);
+    if (queueSize === 0) return;
 
-    let folderName = `${album_artist}/${album.name}`;
-    let tempFileName = `${name} - ${album_artist}.mp3`;
-    let finalFileName = path.join(folderName, `${track_position}- ${name}.mp3`);
+    this.isQueueDownloading = true;
 
-    console.log(
-      `\nTrying to download ${name} - ${album_artist} - ${album.name}`
+    const songToDownload = await this.database.getNextQueueElement();
+    console.log("songToDownload", songToDownload);
+    if (!songToDownload) {
+      console.log("La file d'attente est vide.");
+      this.isQueueDownloading = false;
+      return;
+    }
+    try {
+
+      const downloadingSong = await this.downloadFromTrackId(songToDownload.id);
+      if (downloadingSong.error) {
+        throw new Error("Error while downloading " + downloadingSong.track);
+      }
+
+      await this.database.updateQueueStatus(
+        songToDownload.id,
+        this.database.QUEUE_STATUS.COMPLETED
+      );
+
+    } catch (error) {
+      console.error(
+        "Une erreur s'est produite lors du téléchargement de la chanson :",
+        error
+      );
+      await this.database.updateQueueStatus(
+        songToDownload.id,
+        this.database.QUEUE_STATUS.ERROR
+      );
+    }
+    this.isQueueDownloading = false;
+    this.downloadNextSongInQueue();
+  }
+
+  async downloadVerifQueue(track_data) {
+    const downloadQueueSize = await this.database.getPendingQueueSize();
+
+    let finalFileName = path.join(
+      `${track_data.album_artist}/${track_data.album.name}`,
+      `${track_data.track_position}- ${track_data.name}.mp3`
     );
 
     if (fs.existsSync(path.join(this.FINAL_PATH, finalFileName))) {
@@ -311,29 +328,87 @@ class MusicFunctions {
         path: finalFileName,
       };
     }
-    
-    await this.downloader.downloadTrack(track_data, tempFileName);
-    
-    await utils.ensureDir(path.join(this.FINAL_PATH, folderName));
 
-    await this.downloadCoverTrack(track_data);
-    await this.tagCoverTrack(track_data);
-
-    await this.lyricsController.getAndSaveTxtLyrics(
-      `${name} ${album_artist}`,
-      path.join(this.FINAL_PATH, finalFileName.replace("mp3", "txt"))
+    await this.database.addTrackDataToQueue(
+      track_data,
+      this.database.QUEUE_STATUS.PENDING
     );
 
-    fs.unlinkSync(path.join(this.TEMP_SONGS_PATH, tempFileName));
-    fs.unlinkSync(
-      path.join(this.TEMP_COVERS_PATH, tempFileName.replace("mp3", "jpg"))
-    );
+    // this.downloadNextSongInQueue();
+    if (downloadQueueSize === 0) {
+      return {
+        message: "Track added to download queue",
+        status: "downloading",
+      };
+    } else {
 
-    console.log(`Sucessfully downloaded ${tempFileName}\n`);
+        return {
+          message: "Track added to download queue",
+          status: "pending",
+        };
+      }
+  }
 
-    return {
-      path: finalFileName,
-    };
+  async downloadFromDatas(track_data) {
+    return new Promise(async (resolve, reject) => {
+      // Download a song from Spotify using the data object returned by the search function
+      let { name, album_artist, album, track_position } = track_data;
+
+      let folderName = `${album_artist}/${album.name}`;
+      let tempFileName = `${name} - ${album_artist}.mp3`;
+      let finalFileName = path.join(
+        folderName,
+        `${track_position}- ${name}.mp3`
+      );
+
+      console.log(
+        `\nTrying to download ${name} - ${album_artist} - ${album.name}`
+      );
+
+      if (fs.existsSync(path.join(this.FINAL_PATH, finalFileName))) {
+        console.log(`File ${finalFileName} already exists`);
+        return {
+          message: "File already exists",
+          path: finalFileName,
+        };
+      }
+
+      try {
+        await this.downloader.downloadTrack(track_data, tempFileName);
+  
+        await utils.ensureDir(path.join(this.FINAL_PATH, folderName));
+  
+        await this.downloadCoverTrack(track_data);
+        await this.tagCoverTrack(track_data);
+  
+        await this.lyricsController.getAndSaveTxtLyrics(
+          `${name} ${album_artist}`,
+          path.join(this.FINAL_PATH, finalFileName.replace("mp3", "txt"))
+        );
+  
+        fs.unlinkSync(path.join(this.TEMP_SONGS_PATH, tempFileName));
+        fs.unlinkSync(
+          path.join(this.TEMP_COVERS_PATH, tempFileName.replace("mp3", "jpg"))
+        );
+  
+        await this.database.updateQueueStatus(
+          track_data.id,
+          this.database.QUEUE_STATUS.COMPLETED
+        );
+  
+        console.log(`Sucessfully downloaded ${tempFileName}\n`);
+
+        resolve({
+          message: "File downloaded",
+          error: false,
+          path: finalFileName,
+        })
+        
+      } catch (error) {
+        console.error("Error while downloading", error);
+        reject(error);
+      }
+    });
   }
 
   async downloadFromTrackId(track_id) {
@@ -381,7 +456,16 @@ class MusicFunctions {
       },
     };
 
-    return this.downloadFromDatas(track_data);
+    try {
+      const result = await this.downloadFromDatas(track_data);
+      return result
+    } catch (error) {
+      console.error("Error while downloading", error);
+      return {
+        error: true,
+        track: track_data.name + " - " + track_data.artist,
+      }
+    }
   }
 }
